Fix import paths and move closing-remarks-header-tag to components/
